+++ esda/tests/test_ljc.py
@@ -4,27 +4,27 @@ import numpy as np
 from libpysal.weights.util import lat2W
 from libpysal.common import pandas
 
-from ..local_join_count import Local_Join_Count
+from esda.join_count_local import Join_Count_Local
 
 PANDAS_EXTINCT = pandas is None
 
-class Local_Join_Counts_Tester(unittest.TestCase):
+class Join_Count_Locals_Tester(unittest.TestCase):
     """Unit test for Local Join Counts (univariate)"""
     def setUp(self):
         self.w = lat2W(4, 4)
         self.y = np.ones(16)
         self.y[0:8] = 0
 
-    def test_Local_Join_Counts(self):
+    def test_Join_Count_Locals(self):
             """Test method"""
             np.random.seed(12345)
-            ljc = Local_Join_Count(connectivity=self.w).fit(self.y)
+            ljc = Join_Count_Local(connectivity=self.w).fit(self.y)
             assert np.array_equal(ljc.LJC, [0, 0, 0, 0, 0, 0, 0, 0, 2, 3, 3, 2, 2, 3, 3, 2])
             
             
 suite = unittest.TestSuite()
 test_classes = [
-    Local_Join_Counts_Tester
+    Join_Count_Locals_Tester
 ]
 for i in test_classes:
     a = unittest.TestLoader().loadTestsFromTestCase(i)
