+++ esda/tests/test_local_geary.py
@@ -1,14 +1,12 @@
 import unittest
 import libpysal
 from libpysal.common import pandas, RTOL, ATOL
-from esda import local_geary
+from esda.geary_local import Geary_Local
 import numpy as np
 
 PANDAS_EXTINCT = pandas is None
 
-from ..local_geary import Local_Geary
-
-class Local_Geary_Tester(unittest.TestCase):
+class Geary_Local_Tester(unittest.TestCase):
     def setUp(self):
         np.random.seed(10)
         self.w = libpysal.io.open(libpysal.examples.get_path("stl.gal")).read()
@@ -16,13 +14,13 @@ class Local_Geary_Tester(unittest.TestCase):
         self.y = np.array(f.by_col['HR8893'])
 
     def test_local_geary(self):
-        lG = Local_Geary(connectivity=self.w).fit(self.y)
+        lG = Geary_Local(connectivity=self.w).fit(self.y)
         self.assertAlmostEqual(lG.localG[0], 0.696703432)
         self.assertAlmostEqual(lG.p_sim[0], 0.19)
         
 suite = unittest.TestSuite()
 test_classes = [
-    Local_Geary_Tester
+    Geary_Local_Tester
 ]
 for i in test_classes:
     a = unittest.TestLoader().loadTestsFromTestCase(i)
