+++ esda/tests/test_losh.py
@@ -2,12 +2,10 @@
 import unittest
 import libpysal
 from libpysal.common import pandas, RTOL, ATOL
-from esda import losh
+from esda.losh import LOSH 
 import numpy as np
 
 PANDAS_EXTINCT = pandas is None
-
-from ..losh import losh
 
 class Losh_Tester(unittest.TestCase):
     def setUp(self):
@@ -17,7 +15,7 @@ class Losh_Tester(unittest.TestCase):
         self.y = np.array(f.by_col['HR8893'])
 
     def test_losh(self):
-        ls = losh(connectivity=self.w, inference="chi-square").fit(self.y)
+        ls = LOSH(connectivity=self.w, inference="chi-square").fit(self.y)
         self.assertAlmostEqual(ls.Hi[0], 0.77613471)
         self.assertAlmostEqual(ls.pval[0], 0.22802201)
